requirements.txt
google-cloud-vision==2.4.2
boto3==1.18.20
botocore==1.21.20
tqdm==4.62.0

python-dateutil~=2.8.1
Pillow~=8.3.1
